Continued on ServerFile and Person
Created new class Person.java, that will be used to represent each
voucher on a file

Continued work on ServerFile.java, which is used to represent each
file. A lot of work is needed in this file, for calculations of circles
of trust, vouching etc.

Also, I’m using the jgrapht external library for the graph data
structure that I (will) use to represent a circle of trust