More work on circle of trust
A lot more work in ServerFile.java (and yet more work needed till this
stuff is done).

Basically, vouching works properly, and we’ve gotten as far as the
graph’s nodes being set up. Now we need to add the edges to the graph,
(which will also involve validating signatures, a bit of work!), and
then calculating a circle.

Note: It turns out Person.java is completely unnecessary. It’s fine to
represent a person by the cleartext portion of their certificate, since
that contains lots of information (name, country, organisation etc.).

Also, we need to do some logical work as well, in terms of thinking
about how to define certificate names and such on the server. i.e. if a
client uploads “certificate.crt”, the server should really give it a
name like “cert1.crt”, and then tell the client that their certificate
is called “cert1.crt”, so the client knows what name to use when trying
to vouch

Additionally, we need to think about some stuff to do with required
members.