Files are now replaces
Files are now replaced properly when you request to add a file to the
server that has an identical filename to an existing file. The old
file’s object is deleted. When a file is replaced, a new object is
created for that file and added to the array list ‘files’, so there is
no chance of the old file’s trust information accidentally carrying over