Added an oversight in ServerFile.java
I think I was accidentally ignoring signature verification failing when
constructing the circle of trust. Now, if the verification of a
certificate’s signature fails, the certificate owner is simply tossed
out of the graph, and no edge is created

i.e. if A signs B’s certificate, and C claims that B signed their
certificate, but the signature verification on C’s certificate fails,
then C is tossed out of that graph