Wrote scripts to sign certificates
It works, and I’ve confirmed it works by testing it with the server app
(which I got to report a circle of size 2, between two certificates
that I had sign each other).

However, the way it works is confusing and time consuming. I’ll need to
streamline this entire certificate generation process when I have time.
Alternatively, I can just generate all the certificates beforehand, and
have them ready for the demo