Sadness
It’s quite disappointing, but I couldn’t get it working the way I
wanted, and I’ve given up on doing so.

However, it is still an improvement on before. Basically, instead of
just encrypting the csv file with the public key and decrypting with
the private key, the server now generates a random symmetric key every
time it’s told to close, and a random IV to go with it (AES
encryption), and encrypts the file with that key and IV.

Of course, the problem is that those keys just sit in the directory
along with the file……. I was trying to have them encrypted with the
private key in the last few commits, but Java just would’t allow it

Still, this is better than before, as it can handle a large file.
Public-private key encryption is not meant for large data, it’s
primarily for key encryption. As a consequence, once the csv file
become too large, the encryption failed

This solves that problem, at least