Replacing certificates
Now, when a certificate on the server is replaced (i.e. someone uploads
a.crt, then later uploads another a.crt, replacing the first), the new
certificate is not kept on all files’ list of vouchers (it was before,
i.e. after the replacement, a.crt would still stay on all files’
vouchers lists)