Added security with replacing certificates
A client can now only replace a certificate on the server (i.e. upload
a.crt when the server already has a cert named a.crt) if they have the
private key for the cert they’re trying to replace. Otherwise, the
Server will deny the request. This prevents an attacker from removing
all circles on the server by just uploading conflicting certificates,
at the cost of assuming everyone user has a unique name.

As the original owner of a certificate is now the only person that can
replace it, there is no longer any need to remove the new certificate
from files’ voucher lists, as the owner hasn’t changed.